feat: add initial create custom view modal
